requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
web3==6.11.1
eth-utils==2.3.0
eth-typing==3.5.0
eth-account==0.8.0
